refactor: Consolidate three lipsync pages into unified Lipsync Studio
- Create new Lipsync.tsx with tab-based interface for 3 modes:
  - 1 Person (image + audio → talking video)
  - Multi Person (image + masks + audio tracks → multi-person video)
  - Video Lipsync (video + audio with timeline sync)
- Update App.tsx navigation to single "Lipsync Studio" entry
- Update Homepage.tsx with consolidated lipsync app card
- Add localStorage migration for users with old page values
- Remove separate MultiTalkOnePerson, MultiTalkMultiplePeople, VideoLipsync files
